requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.10

# Validation
pydantic>=2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AI Providers
anthropic==0.7.7
# openai==1.6.1  # Uncomment if using OpenAI
# google-generativeai==0.3.2  # Uncomment if using Google Gemini

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
